Создает board controller
